main.py
#!/usr/bin/env python3
"""
DAX Trend-Following Algorithmic Trading System
Main execution entry point for the trading system.

Author: [Your Name]
Created: 2025
Python Version: 3.9+
"""

import logging
import sys
from pathlib import Path
from datetime import datetime
import yaml

# Add src to Python path
sys.path.append(str(Path(__file__).parent / "src"))

# Core imports
import pandas as pd
import numpy as np

# Local imports (will be implemented in subsequent modules)
# from src.data.loader import DataLoader
# from src.data.validator import DataValidator
# from src.features.engineering import FeatureEngineer
# from src.models.regime import MarketRegimeDetector
# from src.strategy.signals import SignalGenerator
# from src.strategy.risk import RiskManager
# from src.backtesting.engine import BacktestEngine


def setup_logging(log_level: str = "INFO") -> None:
    """Configure logging for the trading system."""
    log_dir = Path("logs")
    log_dir.mkdir(exist_ok=True)
    
    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
    log_file = log_dir / f"trading_system_{timestamp}.log"
    
    logging.basicConfig(
        level=getattr(logging, log_level.upper()),
        format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
        handlers=[
            logging.FileHandler(log_file),
            logging.StreamHandler(sys.stdout)
        ]
    )


def load_config(config_path: str = "config/config.yaml") -> dict:
    """Load system configuration from YAML file."""
    try:
        with open(config_path, 'r') as file:
            config = yaml.safe_load(file)
        logging.info(f"Configuration loaded from {config_path}")
        return config
    except FileNotFoundError:
        logging.warning(f"Config file {config_path} not found, using defaults")
        return {}
    except yaml.YAMLError as e:
        logging.error(f"Error parsing config file: {e}")
        return {}


def validate_environment() -> bool:
    """Validate that the runtime environment meets requirements."""
    try:
        # Check Python version
        if sys.version_info < (3, 9):
            logging.error("Python 3.9+ required")
            return False
            
        # Check critical directories exist
        required_dirs = ["data", "data/raw", "data/processed", "config", "logs"]
        for dir_path in required_dirs:
            Path(dir_path).mkdir(parents=True, exist_ok=True)
            
        logging.info("Environment validation passed")
        return True
        
    except Exception as e:
        logging.error(f"Environment validation failed: {e}")
        return False


def main() -> None:
    """Main execution function for the DAX trading system."""
    print("=" * 60)
    print("DAX Trend-Following Algorithmic Trading System")
    print("=" * 60)
    
    # Setup logging
    setup_logging()
    logger = logging.getLogger(__name__)
    
    try:
        # Validate environment
        if not validate_environment():
            logger.error("Environment validation failed. Exiting.")
            sys.exit(1)
            
        # Load configuration
        config = load_config()
        
        # Initialize system components
        logger.info("Initializing trading system components...")
        
        # TODO: Initialize components as they are implemented
        # data_loader = DataLoader(config.get('data', {}))
        # feature_engineer = FeatureEngineer(config.get('features', {}))
        # regime_detector = MarketRegimeDetector(config.get('regime', {}))
        # signal_generator = SignalGenerator(config.get('strategy', {}))
        # risk_manager = RiskManager(config.get('risk', {}))
        # backtest_engine = BacktestEngine(config.get('backtest', {}))
        
        logger.info("System initialized successfully")
        
        # Main execution logic will be implemented here
        logger.info("Starting main execution loop...")
        
        # Placeholder for main system logic
        print("\n[PLACEHOLDER] Main execution logic will be implemented here")
        print("System components ready for development...")
        
        logger.info("Trading system execution completed")
        
    except KeyboardInterrupt:
        logger.info("System interrupted by user")
    except Exception as e:
        logger.error(f"System error: {e}", exc_info=True)
        sys.exit(1)


if __name__ == "__main__":
    main()